fixed table display problem with phones
The problem was that map initialization code could not get the width
of the hero unit because it is hidden on phones. The javascript would
then crash before the table contents were loaded.

Code now checks for visiblity before initializing map. Side effect is
that if the window width is expanded, without refreshing the page, the
hero unit appears, but without a map.